refactor: 코드리뷰 요청사항 반영
Refs: #44
