+++ src/main/java/com/yesul/admin/repository/AdminRepository.java
@@ -7,6 +7,6 @@ import java.util.Optional;
 
 public interface AdminRepository extends JpaRepository<Admin, Long> {
 
-    Optional<Admin> findByLoginId(String loginId);
     Optional<Admin> findById(Long id);
+
 }
