+++ src/main/java/com/yesul/admin/service/UserDetailsServiceImpl.java
@@ -2,7 +2,8 @@ package com.yesul.admin.service;
 
 import com.yesul.admin.model.dto.auth.LoginAdmin;
 import com.yesul.admin.model.entity.Admin;
-import com.yesul.admin.repository.AdminRepository;
+import com.yesul.exception.handler.AdminNotFoundException;
+import com.yesul.login.repository.AdminLoginRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
@@ -17,13 +18,13 @@ import java.util.List;
 @Service
 public class UserDetailsServiceImpl implements UserDetailsService {
 
-    private final AdminRepository adminRepository;
+    private final AdminLoginRepository adminLoginRepository;
 
     @Override
     public UserDetails loadUserByUsername(String username) throws UsernameNotFoundException {
 
-        Admin fetchedAdmin = adminRepository.findByLoginId(username)
-                .orElseThrow(() -> new IllegalArgumentException("해당 관리자가 존재하지 않습니다."));
+        Admin fetchedAdmin = adminLoginRepository.findByLoginId(username)
+                .orElseThrow(() -> new AdminNotFoundException("해당 관리자가 존재하지 않습니다."));
 
         List<GrantedAuthority> authorityList =
                 List.of(new SimpleGrantedAuthority("ADMIN"));
