+++ src/main/java/com/yesul/login/controller/AdminLoginController.java
@@ -1,0 +1,14 @@
+package com.yesul.login.controller;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+
+@RequestMapping("/admin")
+@Controller
+public class AdminLoginController {
+
+    @GetMapping("/login")
+    public void loginPage(){}
+
+}
