+++ src/main/java/com/yesul/user/repository/UserRepository.java
@@ -18,7 +18,7 @@ public interface UserRepository extends JpaRepository<User, Long> {
             "u.email, u.nickname, u.birthday, u.status ,u.createdAt) FROM User u")
     List<UserListDto> findAllUserInfoDto();
 
-    @Modifying
+    @Modifying(clearAutomatically = true)
     @Query("UPDATE User u SET u.status = :status WHERE u.id = :userId")
     int updateUserStatus(Long userId, Character status);
 }
